+++ src/main/java/com/group10/gamestoryforge/dao/UserDataRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserDataRepository extends JpaRepository<UserData, Integer> {
+public interface UserDataRepository extends JpaRepository<UserData, Long> {
     UserData findByUsername(String username);
 }
